Remove the /service endpoint
Remove the /service endpoint as it is the only remaining code using the
isAdministrator method of the PermissionsService that will not be
removed by the move to the JWT sessions and there is another more
commonly used method for generating service tokens anyway via command
line. The docs for how to do this have been updated.